+++ CHANGELOG.md
@@ -3,9 +3,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/alexdlaird/java-ngrok/compare/1.6.3...HEAD)
+## [Unreleased](https://github.com/alexdlaird/java-ngrok/compare/2.0.0...HEAD)
+
+## [2.0.0](https://github.com/alexdlaird/java-ngrok/compare/1.6.3...2.0.0) - TBD
 ### Changed
-- Changed default installer behavior to download `ngrok` v3 when no version is specified.
+- Default installer behavior to download `ngrok` v3 when no version is specified.
 - Documentation updates.
 - Test updates.